Version without enabling/disabling the HMD
I disable my HMD via my power outlet. Because of this I don't need the enabling/disabling the HMD feature so I have removed it from the script (technically I disabled it rather then removed it). Furthermore SteamVR does not restart anymore when starting the first time because the HMD is already plugged it and activated before launching SteamVR via the Indes Controllers.